chore(evm): cleanup EVM files (#1582)
* chore(evm): cleanup EVM files

* move EVMConfig
